backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "langchain-community>=0.3.20",
    "langchain-google-genai>=2.1.1",
    "langchain-neo4j>=0.4.0",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
]
